requirements.txt
aiogram==3.17
python-dotenv==1.0.0
apscheduler==3.10.4
aiofiles==23.2.1
asyncpg==0.29.0
pymongo==4.6.3
pymysql==1.1.0
paramiko==3.4.0
asyncssh==2.21.1 